requirements.txt
opencv-python
opencv-contrib-python
prettytable
tqdm
yacs
torchfile
plyfile
munkres
chumpy
trimesh
shyaml
PyYAML>=5.1.2
numpy-quaternion
pygame
keyboard
transforms3d
bvhtoolbox
vedo
imgaug
lap
tensorboard
smplx
pycocotools
plotly
open3d
addict
loguru 